+++ compass/core/hierarchy.py
@@ -5,7 +5,7 @@ from pathlib import Path
 from typing import Generator, Iterable, Literal, Optional, Union
 
 from compass.core._scrapers.hierarchy import HierarchyScraper
-from compass.core.logging import logger
+from compass.core.logger import logger
 from compass.core.logon import Logon
 from compass.core.schemas import hierarchy as schema
