+++ compass/core/utility.py
@@ -9,7 +9,7 @@ from typing import Any, Optional, Union
 import certifi
 import requests
 
-from compass.core.logging import logger
+from compass.core.logger import logger
 
 
 def setup_tls_certs() -> None:
